Separate site into View and Edit areas
- Lock editing to backend
- Disable functionality and appearance of all inputs when in View mode

+++ cedar_root/cedar_core/views.py
@@ -25,7 +25,7 @@ import re
 #from django.views.decorators.csrf import csrf_protect
 
 #@csrf_protect
-def view_references(request):
+def view_references(request, view_type):
     
     refs_list = reference.objects.order_by('key_bibtex')
     
@@ -66,10 +66,10 @@ def view_references(request):
         return render(request, 'cedar_core/view_refs_new.html', context)
     """
     
-    context = {'refs_list': refs_list, 'page_title': 'References'}
+    context = {'refs_list': refs_list, 'page_title': 'References', 'view_type': view_type}
     return render(request, 'cedar_core/view_refs_new.html', context)
 
-def view_factors(request, ref_id):
+def view_factors(request, ref_id, view_type):
     
     try:
         ref = reference.objects.get(pk=ref_id)
@@ -107,6 +107,7 @@ def view_factors(request, ref_id):
         'ref': ref,
         'ref_factors': ref_factors,
         'page_title': 'View Factors',
+        'view_type': view_type,
     }
     return render(request, 'cedar_core/view_factors.html', context)
 
@@ -285,48 +286,68 @@ def export_query(request):
     context = {'query_form': query_form, 'page_title': 'Export a Query'}
     return render(request, 'cedar_core/export_query.html', context)
 
-def expand_factor(request, ref_id, fac_id):
+def expand_factor(request, ref_id, fac_id, view_type):
     
     ref = reference.objects.get(pk=ref_id)
     ref_factors = ref.factor_set.all()
     main_fac = factor.objects.get(id=fac_id)
     result_type = factor._meta.get_field('moa_type_id').value_from_object(main_fac)
     
-    # Create the appropriate factor data form depending on the result type. TO DO: account for Relative Risk
-    if result_type == 4:
-        exp_form = OddsTableForm(initial=model_to_dict(main_fac), instance=main_fac)
-    elif result_type == 2:
-        exp_form = PrevTableForm(initial=model_to_dict(main_fac), instance=main_fac)
+    if request.method == 'POST' and view_type == 'edit':
+        # Create the appropriate factor data form depending on the result type. TO DO: account for Relative Risk
+        if result_type == 4:
+            exp_form = OddsTableForm(request.POST, initial=model_to_dict(main_fac), instance=main_fac)
+        elif result_type == 2:
+            exp_form = PrevTableForm(request.POST, initial=model_to_dict(main_fac), instance=main_fac)
+        else:
+            exp_form = ConTableForm(request.POST, initial=model_to_dict(main_fac), instance=main_fac)
+        
+        if exp_form.is_valid():
+            
+            print('CLEANED DATA')
+            print(exp_form.cleaned_data)
+            
+            output = exp_form.save()
+
     else:
-        exp_form = ConTableForm(initial=model_to_dict(main_fac), instance=main_fac)
+        # Create the appropriate factor data form depending on the result type. TO DO: account for Relative Risk
+        if result_type == 4:
+            exp_form = OddsTableForm(initial=model_to_dict(main_fac), instance=main_fac)
+        elif result_type == 2:
+            exp_form = PrevTableForm(initial=model_to_dict(main_fac), instance=main_fac)
+        else:
+            exp_form = ConTableForm(initial=model_to_dict(main_fac), instance=main_fac)
     
     context = {
         'ref': ref,
         'ref_factors': ref_factors,
         'exp_form': exp_form,
         'page_title': 'View Factors',
+        'view_type': view_type,
     }
     return render(request, 'cedar_core/view_factors.html', context)
 
-def delete_factor(request, ref_id, fac_id):
+def delete_factor(request, ref_id, fac_id, view_type):
     
-    # Delete the factor
-    del_fac = factor.objects.get(id=fac_id)
-    del_fac.delete()
+    if view_type == 'edit':
+        # Delete the factor
+        del_fac = factor.objects.get(id=fac_id)
+        del_fac.delete()
     
-    # Reload the view factors page
-    ref = reference.objects.get(pk=ref_id)
-    ref_factors = ref.factor_set.all()
+        # Reload the view factors page
+        ref = reference.objects.get(pk=ref_id)
+        ref_factors = ref.factor_set.all()
     
-    context = {
-        'ref': ref,
-        'ref_factors': ref_factors,
-    }
+        context = {
+            'ref': ref,
+            'ref_factors': ref_factors,
+            'view_type': view_type,
+        }
     
-    return redirect('/cedar_core/references/' + str(ref_id) + '/factors/')
+        return redirect('/cedar_core/references/' + str(ref_id) + '/factors/' + view_type + '/')
     #return render(request,'cedar_core/view_factors.html', context)
 
-def ref_detail(request, ref_id):
+def ref_detail(request, ref_id, view_type):
 
     try:
         ref = reference.objects.get(pk=ref_id)
@@ -339,7 +360,7 @@ def ref_detail(request, ref_id):
     #init_vals = {k: state[k] for k in set(list(state.keys())) - set(exclude_key)}
     
     # If this is a POST request, process the form data
-    if request.method == 'POST':
+    if request.method == 'POST' and view_type == 'edit':
         print('POST ENTERED')
         
         # Create form instances and populate them with data from the request
@@ -434,30 +455,32 @@ def ref_detail(request, ref_id):
                'note_formset': note_formset,
                'note_helper': note_helper,
                'page_title': 'Update a Reference',
+               'view_type': view_type,
                }
     return render(request, 'cedar_core/ref_detail.html', context)
 
 # Add either a new reference note, or a new location
-def add_ref_info(request, ref_id, form_type):
+def add_ref_info(request, ref_id, form_type, view_type):
     
-    # Get reference object
-    ref = reference.objects.get(pk=ref_id)
+    if view_type == 'edit':
+        # Get reference object
+        ref = reference.objects.get(pk=ref_id)
     
-    # Create a new object linked to this reference
-    #new_loc = location_join.objects.create(reference = ref)
-    if form_type == 'loc':
-        new_obj = location_join(loc_ref_id = ref)
-        redir_path = '/cedar_core/references/' + str(ref_id) + '/#loc-md'
-    else:
-        new_obj = reference_note(loc_ref_id = ref)
-        redir_path = '/cedar_core/references/' + str(ref_id) + '/#notes-md'
+        # Create a new object linked to this reference
+        #new_loc = location_join.objects.create(reference = ref)
+        if form_type == 'loc':
+            new_obj = location_join(loc_ref_id = ref)
+            redir_path = '/cedar_core/references/' + str(ref_id) + '/' + view_type + '/#loc-md'
+        else:
+            new_obj = reference_note(loc_ref_id = ref)
+            redir_path = '/cedar_core/references/' + str(ref_id) + '/' + view_type + '/#notes-md'
     
-    # Save new object and redirect
-    new_obj.save()
+        # Save new object and redirect
+        new_obj.save()
     
-    return redirect(redir_path)
+        return redirect(redir_path)
 
-def factor_detail(request, ref_id, fac_id):
+def factor_detail(request, ref_id, fac_id, view_type):
     
     ref = reference.objects.get(pk=ref_id)
     
@@ -474,7 +497,7 @@ def factor_detail(request, ref_id, fac_id):
             #curr_factor.fields[key].disabled = True
         ##factor_forms[f].fields['factor_title'].disabled = True
         
-    if request.method == 'POST':
+    if request.method == 'POST' and view_type == 'edit':
         fac_form = FactorForm(request.POST, initial=model_to_dict(fac), instance=fac)
         
         if fac_form.is_valid():
@@ -490,6 +513,7 @@ def factor_detail(request, ref_id, fac_id):
     context = {'fac': fac,
                'fac_form': fac_form,
                'page_title': 'Edit Factor',
+               'view_type': view_type,
     }
     return render(request, 'cedar_core/factor_detail_new.html', context)
